+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/OutpostNetherStructure.java
@@ -3,7 +3,6 @@ package com.telepathicgrunt.repurposedstructures.world.structures;
 import com.google.common.collect.Lists;
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.RSStructures;
-import com.telepathicgrunt.repurposedstructures.world.structures.pieces.GeneralJigsawGenerator;
 import net.minecraft.entity.EntityType;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
@@ -14,7 +13,10 @@ import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.MobSpawnInfo;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
+import net.minecraft.world.gen.feature.jigsaw.JigsawManager;
+import net.minecraft.world.gen.feature.structure.AbstractVillagePiece;
 import net.minecraft.world.gen.feature.structure.Structure;
+import net.minecraft.world.gen.feature.structure.VillageConfig;
 import net.minecraft.world.gen.feature.template.TemplateManager;
 
 import java.util.List;
@@ -48,7 +50,19 @@ public class OutpostNetherStructure extends Structure<NoFeatureConfig> {
 
         public void init(DynamicRegistries dynamicRegistryManager, ChunkGenerator chunkGenerator, TemplateManager structureManager, int x, int z, Biome biome, NoFeatureConfig NoFeatureConfig) {
             BlockPos blockPos = new BlockPos(x * 16, 0, z * 16);
-            GeneralJigsawGenerator.addPieces(dynamicRegistryManager, chunkGenerator, structureManager, blockPos, this.components, this.rand, dynamicRegistryManager.get(Registry.TEMPLATE_POOL_WORLDGEN).getOrDefault(START_POOL), 11);
+            JigsawManager.method_30419(
+                    dynamicRegistryManager,
+                    new VillageConfig(() -> dynamicRegistryManager.get(
+                            Registry.TEMPLATE_POOL_WORLDGEN).getOrDefault(START_POOL),
+                            11),
+                    AbstractVillagePiece::new,
+                    chunkGenerator,
+                    structureManager,
+                    blockPos,
+                    this.components,
+                    this.rand,
+                    true,
+                    false);
             this.recalculateStructureSize();
 
             BlockPos lowestLandPos = getHighestLand(chunkGenerator);