+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftEndStructure.java
@@ -4,6 +4,7 @@ import com.google.common.collect.Lists;
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.RepurposedStructures;
 import com.telepathicgrunt.repurposedstructures.world.structures.pieces.RSMineshaftPieces;
+import net.minecraft.client.Minecraft;
 import net.minecraft.entity.EntityType;
 import net.minecraft.util.SharedSeedRandom;
 import net.minecraft.util.math.ChunkPos;
@@ -45,6 +46,7 @@ public class RSMineshaftEndStructure extends RSMineshaftStructure {
                 return RepurposedStructures.RSMineshaftsConfig.barrensIslandsEndMineshafts.get() || landHeight > 20;
             }
         }
+
         return false;
     }
 }